backend: add error handler middleware
Looks status codes, and renders a proper error page according to them.

It doesn't work while doing CRUD operations on the admin side, because
those routes don't use this middleware. Admins uses curl, postman
when they want to operate.
In the future, I have a plan to introduce admin panel in the website.
This middleware can be used for them too, then.